debug_counters: get rid of array V2
The main reason I introduced ArraySpecV2 instead of adding a field
to ArraySpec was because I feared that having to edit 30 things
across 10+ files would create reasonably large room for error.
Now that it's all written in a functional-ish style in one python
script, I only have to change a couple of lines to add a field everywhere.